Updates and merged changes between Mikro and akadamson a) I fixed a few typos, added some comments and fixed some formatting (no biggies here) b) I left all 4 flight modes selectable, the 3 stables with a #define in aeroquad.pde, and I made the associated changes to get the process and function hooked up.  I did not hook back up the LED, etc on anything but the AQ Stable c) I updated datastorage.h with scaled default values d) Ok, I did update the ACRO mode, ROLL, PITCH and YAW with SI units.  So you'll need to times 50 those values.  I made it so you could switch back and forth if you wanted... I left the old lines in each flight section in flightcontrol.pde, when you flip back you just simple stick your old P,I,D's in for ACRO/YAW, or the new 50 times values.  I can't tell any difference between new values or old, but for future use, it's probably better with SI. e) While I used receiver.xmit factor, I didn't change it's function, you still can set it to what you want and still slows down the effect of transmitter stick input.

+++ AeroQuad.h
@@ -145,7 +145,8 @@ byte flightMode;
 int minAcro; // Read in from EEPROM, defines min throttle during flips
 
 // Auto level setup
-float levelAdjust[2] = {0.0,0.0};
+//float levelAdjust[2] = {0.0,0.0};
+int levelAdjust[2] = {0,0};
 int levelLimit; // Read in from EEPROM
 int levelOff; // Read in from EEPROM
 // Scale to convert 1000-2000 PWM to +/- 45 degrees
